inching forward, putting some code changes back in
