+++ app/feature_flags.py
@@ -22,10 +22,6 @@ class FeatureFlag(Enum):
     V3_ENABLED = 'V3_ENABLED'
     COMP_AND_PEN_MESSAGES_ENABLED = 'COMP_AND_PEN_MESSAGES_ENABLED'
     VA_PROFILE_EMAIL_STATUS_ENABLED = 'VA_PROFILE_EMAIL_STATUS_ENABLED'
-    VA_PROFILE_V3_COMBINE_CONTACT_INFO_AND_PERMISSIONS_LOOKUP = (
-        'VA_PROFILE_V3_COMBINE_CONTACT_INFO_AND_PERMISSIONS_LOOKUP'
-    )
-    VA_PROFILE_V3_IDENTIFY_MOBILE_TELEPHONE_NUMBERS = 'VA_PROFILE_V3_IDENTIFY_MOBILE_TELEPHONE_NUMBERS'
 
 
 def accept_recipient_identifiers_enabled():
